# kratos/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Greeter API
    description: The greeting service definition.
    version: 0.0.1
paths:
    /hello/{name}:
        get:
            tags:
                - Greeter
            description: Hello name endpoint that returns "my name is {name}"
            operationId: Greeter_SayHelloName
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloNameReply'
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
    /ping:
        get:
            tags:
                - Greeter
            description: Ping endpoint that returns pong
            operationId: Greeter_Ping
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.PingReply'
components:
    schemas:
        helloworld.v1.HelloNameReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the name response
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        helloworld.v1.PingReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message for ping
tags:
    - name: Greeter